Move files to the correct package
