+++ detekt-generator/src/main/kotlin/io/gitlab/arturbosch/detekt/generator/Generator.kt
@@ -3,7 +3,6 @@ package io.gitlab.arturbosch.detekt.generator
 import io.github.detekt.parser.KtCompiler
 import io.gitlab.arturbosch.detekt.generator.collection.DetektCollector
 import io.gitlab.arturbosch.detekt.generator.printer.CliOptionsPrinter
-import io.gitlab.arturbosch.detekt.generator.printer.DetektPrinter
 import org.jetbrains.kotlin.psi.KtFile
 import java.io.PrintStream
 import java.nio.file.Files
