+++ detekt-generator/src/main/kotlin/io/gitlab/arturbosch/detekt/generator/collection/RuleSetPage.kt
@@ -1,0 +1,6 @@
+package io.gitlab.arturbosch.detekt.generator.collection
+
+data class RuleSetPage(
+    val ruleSet: RuleSetProvider,
+    val rules: List<Rule>
+)
