+++ detekt-generator/src/test/kotlin/io/gitlab/arturbosch/detekt/generator/printer/RuleSetPagePrinterSpec.kt
@@ -1,7 +1,6 @@
 package io.gitlab.arturbosch.detekt.generator.printer
 
 import io.github.detekt.test.utils.resource
-import io.gitlab.arturbosch.detekt.generator.printer.rulesetpage.RuleSetPagePrinter
 import io.gitlab.arturbosch.detekt.generator.util.createRuleSetPage
 import org.assertj.core.api.Assertions.assertThat
 import org.spekframework.spek2.Spek
